fix: declared roleId with wrong datatype

+++ backend/src/main/java/com/feis/eduhub/backend/features/account/Account.java
@@ -15,15 +15,15 @@ public class Account {
     @NonNull
     private String lastName;
     private String icon;
-    private int roleId;
+    private long roleId;
 
-    public Account(String firstName, String lastName, int roleId) {
+    public Account(String firstName, String lastName, long roleId) {
         this.firstName = firstName;
         this.lastName = lastName;
         this.roleId = roleId;
     }
 
-    public Account(String firstName, String lastName, String icon, int roleId) {
+    public Account(String firstName, String lastName, String icon, long roleId) {
         this.firstName = firstName;
         this.lastName = lastName;
         this.icon = icon;
